fix: update thumbnail invariably while resetting focus for different cases
